+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.4.1] 12-07-2023
+
+### Added
+
+- Test e2e de error en División por 0
+- Test e2e de error en Potencia por 200000
+
 ## [1.4.0] 17-06-2023
 
 ### Added
@@ -72,7 +79,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Base de la API
 - Funciones core para realizar suma, resta, multiplicación, división y potencia
 
-[unreleased]: https://github.com/Nevy48/recalc/compare/v1.4.0...HEAD
+[unreleased]: https://github.com/Nevy48/recalc/compare/v1.4.1...HEAD
+[1.4.1]:https://github.com/Nevy48/recalc/releases/tag/v1.4.1
 [1.4.0]: https://github.com/Nevy48/recalc/releases/tag/v1.4.0
 [1.3.2]: https://github.com/frlp-utn-ingsoft/recalc/releases/tag/v1.3.2
 [1.3.1]: https://github.com/frlp-utn-ingsoft/recalc/releases/tag/v1.3.1